python-order-service/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
certifi==2026.4.22
cffi==2.0.0
click==8.3.3
cryptography==47.0.0
fastapi==0.136.1
h11==0.16.0
hiredis==3.3.1
httpcore==1.0.9
httpx==0.28.1
idna==3.13
jsonpath-ng==1.8.0
ml_dtypes==0.5.4
more-itertools==10.8.0
numpy==2.4.4
pycparser==3.0
pydantic==2.13.3
pydantic-extra-types==2.11.1
pydantic_core==2.46.3
python-dotenv==1.2.2
python-ulid==3.1.0
PyYAML==6.0.3
redis==7.4.0
redis-om==1.1.0
redisvl==0.18.0
setuptools==82.0.1
starlette==1.0.0
tenacity==9.1.4
types-cffi==2.0.0.20260408
types-pyOpenSSL==24.1.0.20240722
types-redis==4.6.0.20241004
types-setuptools==82.0.0.20260408
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
